+++ hibernate/src/main/java/ru/origami/hibernate/attachment/HibernateAttachment.java
@@ -20,10 +20,10 @@ import static ru.origami.testit_allure.test_it.testit.aspects.StepAspect.TEST_IT
 public class HibernateAttachment {
 
     @Attachment(value = "sql.query")
-    public static byte[] attachSqlQueryToAllure(String queryString, List<QueryParameter> parameters) {
+    public static byte[] attachSqlQueryToAllure(String queryString, List<QueryParameter> parameters, String schema) {
         String query = Objects.requireNonNull(getSqlQuery(queryString, parameters), getLangValue("hibernate.null.attachment"));
         attachQueryToTestIT(query);
-        showSql(query);
+        showSql(query, schema);
 
         return query.getBytes();
     }
@@ -58,9 +58,14 @@ public class HibernateAttachment {
         return stackTrace;
     }
 
-    public static void showSql(String query) {
+    public static void showSql(String query, String schema) {
         if (isLocal() || isLoggingEnabled()) {
-            log.info("{}:\n{}\n", getLangValue("hibernate.query.word"), query);
+            if (Objects.nonNull(schema)) {
+                log.info("{}: {}. {}:\n{}\n", getLangValue("hibernate.query.schema"), schema,
+                        getLangValue("hibernate.query.word"), query);
+            } else {
+                log.info("{}:\n{}\n", getLangValue("hibernate.query.word"), query);
+            }
         }
     }