+++ hibernate/src/main/java/ru/origami/hibernate/queries/DBQuery.java
@@ -7,11 +7,11 @@ import ru.origami.hibernate.models.EHibernateResource;
 @Slf4j
 public class DBQuery<R> extends DBCommonQuery<DBQuery<R>> {
 
-    public DBQuery(Session session, String query, EHibernateResource resource) {
-        super(session, query, resource);
+    public DBQuery(Session session, String query, EHibernateResource resource, String schema) {
+        super(session, query, resource, schema);
     }
 
-    public DBQuery(Session session, String query, Class<?> resultType, EHibernateResource resource) {
-        super(session, query, resultType, resource);
+    public DBQuery(Session session, String query, Class<?> resultType, EHibernateResource resource, String schema) {
+        super(session, query, resultType, resource, schema);
     }
 }